1. moved all classes to new package BJS2_68765 2. class ProductManager, updated method groupProductsByCategory
