+++ src/main/java/school/faang/BJS2_68765/Category.java
@@ -1,4 +1,4 @@
-package school.faang;
+package school.faang.BJS2_68765;
 
 public enum Category {
     FOOD, ELECTRONICS, CLOTHING, OTHER
